Move configuration to custom folder
